refactor: grpc wrapper
